Update breakout.c
The following is the code necessary to make the classic arcade game "Breakout." The basic framework for the game was given through Harvard's online CS50 course. I contributed the code necessary to create the bricks, the ball, the paddle and the scoreboard, ("initBricks," "initBall," "initPaddle," and "initScoreboard," respectively,) as well as the code involved in moving (and in the case of bricks, removing) these objects to make the game functional.